+++ common/src/main/java/com/telepathicgrunt/the_bumblezone/blocks/EssenceBlockBlue.java
@@ -60,7 +60,7 @@ public class EssenceBlockBlue extends EssenceBlock {
 
     @Override
     public int getEventTimeFrame() {
-        return 10800;
+        return BzGeneralConfigs.blueArenaTimeFrameInTicks;
     }
 
     @Override