+++ common/src/main/java/com/telepathicgrunt/the_bumblezone/blocks/EssenceBlockGreen.java
@@ -4,6 +4,7 @@ import com.telepathicgrunt.the_bumblezone.Bumblezone;
 import com.telepathicgrunt.the_bumblezone.blocks.blockentities.EssenceBlockEntity;
 import com.telepathicgrunt.the_bumblezone.bossbars.ServerEssenceEvent;
 import com.telepathicgrunt.the_bumblezone.client.rendering.rootmin.RootminPose;
+import com.telepathicgrunt.the_bumblezone.configs.BzGeneralConfigs;
 import com.telepathicgrunt.the_bumblezone.entities.mobs.RootminEntity;
 import com.telepathicgrunt.the_bumblezone.modinit.BzEntities;
 import com.telepathicgrunt.the_bumblezone.modinit.BzSounds;
@@ -52,7 +53,7 @@ public class EssenceBlockGreen extends EssenceBlock {
 
     @Override
     public int getEventTimeFrame() {
-        return 6000;
+        return BzGeneralConfigs.greenArenaTimeFrameInTicks;
     }
 
     @Override