+++ common/src/main/java/com/telepathicgrunt/the_bumblezone/blocks/EssenceBlockPurple.java
@@ -3,6 +3,7 @@ package com.telepathicgrunt.the_bumblezone.blocks;
 import com.telepathicgrunt.the_bumblezone.Bumblezone;
 import com.telepathicgrunt.the_bumblezone.blocks.blockentities.EssenceBlockEntity;
 import com.telepathicgrunt.the_bumblezone.bossbars.ServerEssenceEvent;
+import com.telepathicgrunt.the_bumblezone.configs.BzGeneralConfigs;
 import com.telepathicgrunt.the_bumblezone.entities.nonliving.PurpleSpikeEntity;
 import com.telepathicgrunt.the_bumblezone.modinit.BzBlocks;
 import com.telepathicgrunt.the_bumblezone.modinit.BzEntities;
@@ -42,7 +43,7 @@ public class EssenceBlockPurple extends EssenceBlock {
 
     @Override
     public int getEventTimeFrame() {
-        return 5020;
+        return BzGeneralConfigs.purpleArenaTimeFrameInTicks;
     }
 
     @Override